add file class and extend directory
